No longer the AI's job to switch reward values.
The AI just needs to know what are the A and B values for all rewards.
It will be up to the calling code (UI) to handle which team values are
put in team A or team B. It will instantiate a new DraftAI object for
this and optionally pass in a save TT for it to load.